feat(authentication-service): add config and a new otp controller
