+++ services/authentication-service/src/enums/mfa-config.enum.ts
@@ -1,4 +1,4 @@
-export const enum SecondFactor {
+export const enum OtpMethodType {
   OTP = 'OTP',
   GOOGLE_AUTHENTICATOR = 'Google Authenticator App',
 }
