+++ services/authentication-service/src/types.ts
@@ -1,7 +1,8 @@
 /* eslint-disable  @typescript-eslint/naming-convention */
 import {AnyObject} from '@loopback/repository';
 import {IServiceConfig} from '@sourceloop/core';
-import {SecondFactor} from './enums';
+import {STRATEGY} from 'loopback4-authentication';
+import {OtpMethodType} from './enums';
 import {LocalUserProfileDto} from './models';
 import {SignupRequestDto} from './models/signup-request-dto.model';
 
@@ -13,7 +14,11 @@ export const AuthDbSourceName = 'AuthDB';
 export const AuthCacheSourceName = 'AuthCache';
 
 export interface IMfaConfig {
-  secondFactor: SecondFactor;
+  secondFactor: STRATEGY;
+}
+
+export interface IOtpConfig {
+  method: OtpMethodType;
 }
 
 export interface PreSignupFn<T, S> {
